refactor: move helpers inside Permission folder
